Updated releng/autorelease Project: ovsdb  533471b20e93fcf7790f17b5cca80a89694f4255
For controller initiated connection, there is race condition
while disconnecting the device from controller, that does
not allow one of the non-owner southbound instance to disconnect
the device from the plugin. Race condition was occuring because
OvsdbConnectionInstance was not stored in the internal cache
at correct point in control flow.It gets added when EntityOwnershipChange
event happens, so when data change event for disconnection occured,
it requires OvsdbConnectionInstance to be in cache to disconnect
from the device, but OvsdbConnectionInstance was not in the cache
becuse EntityOwnershipService event was not recieved by that non-owner instance.

Change-Id: I3861b5f2907d08709a3bd1aa4f4ecfbfd444351d
Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>