Adding scripts to generate release notes for Service Release.
The scripts will generate the list of changes between two
releases for all projects in both wiki and asciidocs format.

The first phase is an initial pass based on existing scripts
with minimal changes. In the second phase, we should refactor
the scripts to 1) execute with minimal parameters, 2) autogenerate
the list of projects depending on the branch by reading
from pom, 3) format the wiki and ascii outputs to be ready-to-push
complete pages, and 4) executed by a Jenkins job.

Change-Id: Ic86c3c69385dc3be4a2c279c4577e75c6f8ce3a1
Signed-off-by: A H <an.ho@huawei.com>