Updated releng/autorelease Project: bgpcep  8133f9ffe21c405688228d7f33de8ec32f690a9d
BUG-3771: Use sync on FutureChannels intead of await

"Await" waits until future has finished and throw execution
if thread was interrupted."Sync"" instead waits for this
future until it is done, and re throws the cause of
the failure if this future failed.
When closing the module, first we ensure that channel
has been closed and if it didn't fail we proceed
with next closures.

Change-Id: I1240d6fdb6986b8c30251206d8027db7b1a987fa
Signed-off-by: Claudio D. Gasparini <cgaspari@cisco.com>