Updated releng/autorelease Project: bgpcep  04633daf114ccb6e5a8e83f39229d3dc3f41ab36
BUG-4634: Prefix in withdrawn list is not ignored when announced in same msgUpd.

An UPDATE message SHOULD NOT include the same address prefix in the
WITHDRAWN ROUTES and Network Layer Reachability Information fields.
However, a BGP speaker MUST be able to process UPDATE messages in
this form.  A BGP speaker SHOULD treat an UPDATE message of this form
as though the WITHDRAWN ROUTES do not contain the address prefix.
Fix by ingore common NLRI on withdrawn router

Change-Id: Ic9ebd3ce13f042f0dfcd2b28815442f8484add16
Signed-off-by: Claudio D. Gasparini <cgaspari@cisco.com>