Updated releng/autorelease Project: netconf  22917d7df75d6d97a62d1c32a4f8901097e6c5ad
BUG-4638: use interface instead of concrete types

The parser is bound to switch its implementation of TypeDefinitions. As
a first step, look at the implemented type interface instead of looking
for a concrete implementation class.

Change-Id: I86b3b03d30e9098cbbd99bc61c23bc859ef02302
Signed-off-by: Robert Varga <rovarga@cisco.com>

BUG-4638: handle StringTypeDefinition properly

This patch adds the path for StringTypeDefinition, so we do not lose the
length restrictions. I do not believe their mapping is accurate, though.

Change-Id: If400f97f9440dab5e5f77cdc2c1eaddcdb2979f8
Signed-off-by: Robert Varga <rovarga@cisco.com>