Updated releng/autorelease Project: centinel  1198af3e31a48b16c9491aecf75f2da09e6732ef
Fix relativePath to parent pom and cleanup

<relativePath/> causes Maven to be unable to find the parent pom since
this sets the relative path to the parent to null. The true relative
path to the parent is one directory above which is the default value of
relative path so we can remove this entry.

Also cleanup the parent section removing the unnecessary comments.

Change-Id: Ia3cd6ae2d9938b27d7b7e28b82a88c608a3f43bd
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>