Updated releng/autorelease Project: yangtools  976487b65036c293658ed17f18b44cae359bab2d
BUG-4684: prune WRITE operation children when ready

When a DataTreeModification is ready, as part of the seal operation, we
squash any child operations into the written value.

This shifts the cost of performing this operation from the commit path
to the user path. It also prevents interactions with child operations
blowing up in min/max validation.

The validation problem remains if the parent operation is MERGE, as that
case requires current data tree state to be present. A subsequent patch
will provide a complete fix there.

Change-Id: I9c96a10fdf5301fc72cd67cefe222fac527a5673
Signed-off-by: Robert Varga <rovarga@cisco.com>