Updated releng/autorelease Project: ovsdb  fbcc59d5065e49650ce64109e863a796f314617e
Avoid exceptions with IPv6 subnets

In a setup where ODL is used for L2 support and Neutron L3 agent
is used for Routing, we see some Java exceptions when IPv6 subnets
are created. Because of the Java exception, Layer2 connectivity
is lost and we are unable to work with IPv6 subnets. In this patch,
we log an error instead of exception.

Patch set 2: add other places in netvirt code where ipv6 (address and prefix)
             could be used and it would be improperly interpreted.

Patch set 3: minor code review changes.

Change-Id: I4e253f8f86a1a3f027e57861549e9a1f98ff5fbf
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>
Signed-off-by: SridharG <sgaddam@redhat.com>