Updated releng/autorelease Project: aaa  02ff8ceebb2101036735f06c3fa730905d0f65ca
Formatting applied to RoleHandler

This change tackles some vital formatting issues with RoleHandler.  The
following is done to the RoleHandler class to promote clarity and conformance
to project style:

1) Move the class documentation comment to the right location.  Add some
detailed information to the class comment concerning how to use the REST
endpoints for RoleHandler.

2) Add function documentation comments explaining their responsibilities.

3) Format in eclipse by clicking "format".

4) Fix any errors reported by "mvn checkstyle:checkstyle".

No actual code or functionality changes were made.  This is strictly a
formatting related patch.

Change-Id: I7a8a9343792d5384b1639010918d2ed8ecdaaa6a
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>