Updated releng/autorelease Project: aaa  e4fa925b0ed179dfc3eb6ac5efa4a8b4456f4bf1
Formatting applied to IdmLightApplication

This change tackles some vital formatting issues with IdmLightApplication.  The
following is done to the IdmLightApplication class to promote clarity and conformance
to project style:

1) Move the class documentation comment to the right location.  Add some
detailed information to the class comment concerning different handlers.

2) Add CONSTANT documentation.

3) Format in eclipse by clicking "format".

4) Fix any errors reported by "mvn checkstyle:checkstyle".

5) Add a TODO to address the mismatch in MAX_FIELD_LENGTH between this
class and the Store classes, which assume 128 not 256.

No actual code or functionality changes were made.  This is strictly a
formatting related patch.

Change-Id: Id4fee54715f392830a97207f24fc9f7ca6880b60
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>