Updated releng/autorelease Project: aaa  3a7678312c31474f04827d541ea69dc09b1a8424
Formatting applied to StoreBuilder

This change tackles some vital formatting issues with StoreBuilder.  The
following is done to the IdmLightApplication class to promote clarity and conformance
to project style:

1) Add some detailed information to the class comment concerning the default installed
accounts.

2) Add TODO concerning what to do if "sdn" is deleted.  Implies that "sdn" could
be made a mandatory domain.

3) Format in eclipse by clicking "format".

4) Fix any errors reported by "mvn checkstyle:checkstyle".

No actual code or functionality changes were made.  This is strictly a
formatting related patch.

Change-Id: I09f45e529201d3eeaca94286f86e3dc06bf727c8
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>