Updated releng/autorelease Project: sxp  c6556358ea3d383fe2b96d81944f75e15c4a93a9
Switch to features-parent and drop ${netty.version}

Clean up the TODOs in features/pom.xml.

Depend on netty-handler instead of netty-all in sxp-core (the code
only needs netty-handler to build).

Depend on individual netty components instead of netty-all in features
so we can use odlparent's dependency management. Drop
netty-codec-socks which shouldn't be necessary (and isn't defined in
odlparent).

Change-Id: I5e212929ef5761cfdff2cdba403830aa40746750
Signed-off-by: Stephen Kitt <skitt@redhat.com>