Updated releng/autorelease Project: integration/distribution  5812af7145b3733a9b4a37b7d0e563e6ec04bcdf
Fix deploy job by removing features from integration

This is a sqashed commit of two changes:
https://git.opendaylight.org/gerrit/31719
https://git.opendaylight.org/gerrit/31715

Both are needed for integration/distribution verify job to pass.
Here are their messages:

Comment out features of DIDM and NIC

This is an attempt to fix deploy test job.
Those features are bringing in odl-openflowplugin-*-li,
which lead to BindExceptions, which lead to red dot.
Feature repositories are still present.

Comment out features of Centinel and Sxp

This is an attempt to fix install:all installation.
Yang model errors are seen in karaf.log for those projects.
Feature repositories are still present.

Change-Id: Ice77a35c27dfc7240aff555dea35a138c2e683a4
Signed-off-by: Vratko Polak <vrpolak@cisco.com>