Updated releng/autorelease Project: yangtools  f29f6b8445fae2505bce3feedfb3d005695adcdd
BUG-4295: instantiate MERGE operations lazily

This patch reworks how merges are done in a DataTreeModification by
moving the logic to SchemaAwareApplyOperation, which is the final
recipient of the resulting ModifiedNode.

This way the code is co-located and can be specialized based on
information available for that particular node, and the container merge
code is cleanly separated from the leaf node code, which turns each
merge into a write.

When a merge occurs on a previously-written node, we graft all merged
children onto the write, using recursion only when necessary.

checkPresentChild method renamed.
Fix of ModifiedNode and added unit test.

Change-Id: I674e3d2150e796472e831abdcfa0fad582b69759
Signed-off-by: Robert Varga <rovarga@cisco.com>
Signed-off-by: Filip.Gregor <fgregor@cisco.com>
Signed-off-by: Peter Kajsa <pkajsa@cisco.com>