Updated releng/autorelease Project: lispflowmapping  3ae3c9d53e99404a77401ac309f79e00644bcb36
Bug 4822: Simplify access to configuration variables

This patch removes the need to pass on configuration variables as part
of the API. Instead, any class can get access to any configuration
variable by accessing a single static instance of ConfigIni. This opens
up the possibility of adding new knobs to modify behavior of parts of
the code, without the need to change APIs (adding a method call
parameter for each knob).

The patch doesn't change the existing APIs (they should be cleaned up in
Boron timeframe though), but allows new confguration options with no API
changes. The main reason for this patch NOW is Bug 4822, which is
related to the merging of locators, and needs support for a new
configuration option.

Change-Id: I0f3a8680c749034f2b784cd5c24379dc3f589538
Signed-off-by: Lorand Jakab <lojakab@cisco.com>