Updated releng/autorelease Project: odlparent  6688af5b7d46c4682ef0724b6466c6f98e80c1ed
Use more specific dependencies than karaf-maven-plugin

PerRepoTestRunner currently relies on karaf-maven-plugin to provide a
URLStreamHandlerFactory. This involves pulling in all
karaf-maven-plugin's dependencies, which end up contributing to all
ODL projects' complex dependencies trees as soon as they start using
features-test.

To avoid this, copy CustomBundleURLStreamHandlerFactory from
karaf-maven-plugin; the code is trivial. This allows
karaf-maven-plugin to be removed from the dependencies, as long as the
specific deployers we need are added. I copied mvn, wrap, feature,
spring and blueprint for now. A number of other dependencies become
unnecessary as a result.

Another approach would be to exclude all the irrelevant dependencies
from karaf-maven-plugin but this feels more fragile to me. With our
own CustomBundleURLStreamHandlerFactory we get to decide what
protocols we want to support.

Change-Id: I6b660582d0fbca57d8f20c7dd53375884a982e10
Signed-off-by: Stephen Kitt <skitt@redhat.com>