Updated releng/autorelease Project: ovsdb  00594d7fc52cc75462d18162a9844138b7601078
Remove failing SouthboundIT test because of new yang enforcement

https://bugs.opendaylight.org/show_bug.cgi?id=4355 added enforcement of values
marked as mandatory in a leaf. The failing tests were negative testing for
setting external_ids with an empty value. That test is invalid now that the
mandatory value is enforced.

Missed a instance in this patch: https://git.opendaylight.org/gerrit/#/c/32321/
but it was already merged.

Change-Id: I5fcc140907b4deec69e025a5055963f4a6743a01
Signed-off-by: Sam Hague <shague@redhat.com>